+++ spec/features/altering_patients_spec.rb
@@ -16,12 +16,35 @@ RSpec.feature 'Patient altering' do
     context 'when registering a new patient' do
       before { stub_webdav }
 
+      scenario 'blocks incorrect case number registration' do
+        visit new_patient_path
+
+        fill_in 'patient[case_number]', with: '[a stranger in the night]'
+
+        expect { click_button I18n.t('register') }.
+          not_to(change { Patient.count })
+
+        expect(page).
+          to have_content I18n.t 'activerecord.errors.models.patient.attributes.case_number.invalid'
+      end
+
       scenario 'lets the user register a case with case number' do
         visit new_patient_path
 
         expect(page).to have_content I18n.t('simple_form.labels.patient.case_number')
 
         fill_in 'patient[case_number]', with: '888'
+
+        expect { click_button I18n.t('register') }.
+          to change { Patient.count }.by(1)
+
+        expect(current_path).to eq patient_path(Patient.first)
+      end
+
+      scenario 'lets the user register a case with uncommon characters' do
+        visit new_patient_path
+
+        fill_in 'patient[case_number]', with: '{*^+-'
 
         expect { click_button I18n.t('register') }.
           to change { Patient.count }.by(1)
@@ -81,6 +104,18 @@ RSpec.feature 'Patient altering' do
 
         expect(current_path).to eq patient_path(Patient.first)
         expect(page).to have_content '1234'
+      end
+
+      scenario 'automatically synchronizes data_files and updates status for strange case number' do
+        visit new_patient_path
+
+        fill_in 'patient[case_number]', with: '{-+'
+
+        expect { click_button I18n.t('register') }.to change { Patient.count }.by(1)
+
+        expect(current_path).to eq patient_path(Patient.first)
+        expect(Patient.first.data_files.count).to eq 1
+        expect(page).to have_content '{-+'
       end
     end
   end